fix(HQM-248): Move new stacks into route-types.ts
