+++ src/app.tsx
@@ -31,8 +31,8 @@ import {Transaction} from '@app/models/transaction';
 import {VariablesBool} from '@app/models/variables-bool';
 import {Wallet} from '@app/models/wallet';
 import {navigator} from '@app/navigator';
+import {OnboardingStackRoutes} from '@app/route-types';
 import {RootStack} from '@app/screens/RootStack';
-import {OnboardingStackRoutes} from '@app/screens/WelcomeStack/OnboardingStack';
 import {AppTheme, ModalType} from '@app/types';
 import {getAppTrackingAuthorizationStatus, sleep} from '@app/utils';
 import {SPLASH_TIMEOUT_MS} from '@app/variables/common';
