+++ src/event-actions/on-app-mnemonic-backup.ts
@@ -1,6 +1,6 @@
 import {Wallet} from '@app/models/wallet';
 import {navigator} from '@app/navigator';
-import {HomeStackRoutes} from '@app/screens/HomeStack';
+import {HomeStackRoutes} from '@app/route-types';
 
 export async function onAppMnemonicBackup(wallet: Wallet) {
   navigator.navigate(HomeStackRoutes.BackupNotification, {wallet});
