+++ src/screens/HomeStack/StakingDelegateStack/staking-delegate-form.tsx
@@ -15,7 +15,7 @@ import {Wallet} from '@app/models/wallet';
 import {
   StakingDelegateStackParamList,
   StakingDelegateStackRoutes,
-} from '@app/screens/HomeStack/StakingDelegateStack';
+} from '@app/route-types';
 import {Balance} from '@app/services/balance';
 import {Cosmos} from '@app/services/cosmos';
 import {FEE_ESTIMATING_TIMEOUT_MS} from '@app/variables/common';
